fix(evm): fix gRPC-only chain-id
